requirements.txt
pip==23.2.1
cvxpy==1.3.2
matplotlib==3.7.3
pytorch-lightning==1.3.1
scikit-image==0.21.0
scikit-learn==1.3.0
torchvision==0.13.1
torchmetrics==0.3.2
Pillow==9.2.0
numpy==1.23.2